requirements.txt
streamlit
openai
python-dotenv
pillow
tiktoken
requests
fastapi
uvicorn
pydantic